+++ src/matchRoute.js
@@ -9,11 +9,11 @@
 
 import { matchPath, matchBasePath } from './matchPath';
 
-function* matchRoute(route, baseUrl, path) {
+function* matchRoute(route, baseUrl, path, parentParams) {
   let match;
 
   if (!route.children) {
-    match = matchPath(route.path, path);
+    match = matchPath(route.path, path, parentParams);
 
     if (match) {
       yield {
@@ -27,7 +27,8 @@ function* matchRoute(route, baseUrl, path) {
   }
 
   if (route.children) {
-    match = matchBasePath(route.path, path);
+    match = matchBasePath(route.path, path, parentParams);
+
     if (match) {
       yield {
         route,
@@ -37,12 +38,13 @@ function* matchRoute(route, baseUrl, path) {
         params: match.params,
       };
 
-      for (const childRoute of route.children) {
+      for (let i = 0; i < route.children.length; i += 1) {
         const newPath = path.substr(match.path.length);
         yield* matchRoute(
-          childRoute,
+          route.children[i],
           baseUrl + (match.path === '/' ? '' : match.path),
-          newPath.startsWith('/') ? newPath : `/${newPath}`
+          newPath.startsWith('/') ? newPath : `/${newPath}`,
+          match.params
         );
       }
     }