+++ CHANGELOG.md
@@ -2,9 +2,34 @@
 
 All notable changes to this project will be documented in this file.
 
-### [Unreleased][unreleased]
+### [v2.0.0] - 2016-10-20
 
-- Add `context.redirect(path)` method to be used from inside route actions (PLANNED)
+- Preserve `context.params` values from the parent route ([#57](https://github.com/kriasoft/universal-router/pull/57))
+- Throws an error if no route found ([#62](https://github.com/kriasoft/universal-router/pull/62))
+- Remove obsolete `context.end()` method ([#60](https://github.com/kriasoft/universal-router/pull/60))
+- Remove obsolete `match` alias for `resolve` function ([#59](https://github.com/kriasoft/universal-router/pull/59))
+- Do not throw an error for malformed URI params ([#54](https://github.com/kriasoft/universal-router/pull/54))
+- Handle `null` the same way as `undefined` ([#51](https://github.com/kriasoft/universal-router/pull/51))
+- Return `null` instead of `undefined` to signal no match ([#51](https://github.com/kriasoft/universal-router/pull/51))
+- Support `context.next()` across multiple routes ([#49](https://github.com/kriasoft/universal-router/pull/49))
+- Sequential execution of asynchronous routes ([#49](https://github.com/kriasoft/universal-router/pull/49))
+- Remove errors handler from core ([#48](https://github.com/kriasoft/universal-router/pull/48))
+- Drop support of node.js v5 and below ([#47](https://github.com/kriasoft/universal-router/pull/47))
+
+### [v1.2.2] - 2016-05-31
+
+- Update UMD build to include missing dependencies ([#33](https://github.com/kriasoft/universal-router/pull/33))
+
+### [v1.2.1] - 2016-05-12
+
+- Rename `match()` to `resolve()`. E.g. `import { resovle } from 'universal-router'`
+- Fix an issue when the router throws an exception when the top-level route doesn't have `children` property
+- Include CommonJS, Harmony Modules, ES5.1 and UMD builds into NPM package
+- Include source maps into NPM package
+
+### [v1.1.0-beta.4] - 2016-04-27
+
+- Fix optional parameters, e.g. `/products/:id?` ([#27](https://github.com/kriasoft/universal-router/pull/27))
 
 ### [v1.1.0-beta.3] - 2016-04-08
 
@@ -39,9 +64,12 @@ All notable changes to this project will be documented in this file.
 
 - Small bug fixes and improvements
 
-[unreleased]: https://github.com/kriasoft/react-starter-kit/compare/v1.1.0-beta.3...HEAD
-[v1.1.0-beta.3]: https://github.com/kriasoft/react-starter-kit/compare/v1.1.0-beta.2...v1.1.0-beta.3
-[v1.1.0-beta.2]: https://github.com/kriasoft/react-starter-kit/compare/v1.1.0-beta.1...v1.1.0-beta.2
-[v1.1.0-beta.1]: https://github.com/kriasoft/react-starter-kit/compare/v1.0.0-beta.1...v1.1.0-beta.1
-[v1.0.0-beta.1]: https://github.com/kriasoft/react-starter-kit/compare/v0.0.7...v1.0.0-beta.1
-[v0.0.7]: https://github.com/kriasoft/react-starter-kit/compare/v0.0.6...v0.0.7
+[v2.0.0]: https://github.com/kriasoft/universal-router/compare/v1.2.2...v2.0.0
+[v1.2.2]: https://github.com/kriasoft/universal-router/compare/v1.2.1...v1.2.2
+[v1.2.1]: https://github.com/kriasoft/universal-router/compare/v1.1.0-beta.4...v1.2.1
+[v1.1.0-beta.4]: https://github.com/kriasoft/universal-router/compare/v1.1.0-beta.3...v1.1.0-beta.4
+[v1.1.0-beta.3]: https://github.com/kriasoft/universal-router/compare/v1.1.0-beta.2...v1.1.0-beta.3
+[v1.1.0-beta.2]: https://github.com/kriasoft/universal-router/compare/v1.1.0-beta.1...v1.1.0-beta.2
+[v1.1.0-beta.1]: https://github.com/kriasoft/universal-router/compare/v1.0.0-beta.1...v1.1.0-beta.1
+[v1.0.0-beta.1]: https://github.com/kriasoft/universal-router/compare/v0.0.7...v1.0.0-beta.1
+[v0.0.7]: https://github.com/kriasoft/universal-router/compare/v0.0.6...v0.0.7